refactor

+++ src/main/java/org/example/shop/api/controller/CartController.java
@@ -35,7 +35,7 @@ public class CartController {
     }
 
     @PostMapping(value = "/{itemId}", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-    public Mono<String> addToCart(@PathVariable int itemId, OrderRequest request, WebSession session) {
+    public Mono<String> addToCart(@PathVariable Long itemId, OrderRequest request, WebSession session) {
         return orderService.updateOrder(itemId, request.getAction(), session.getId()).thenReturn("redirect:/cart/items");
     }
