+++ src/main/java/org/example/shop/api/controller/ShopController.java
@@ -38,7 +38,7 @@ public class ShopController {
     }
 
     @PostMapping(value = "/main/items/{itemId}", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-    public Mono<String> addToCart(@PathVariable int itemId, OrderRequest request, WebSession session) {
+    public Mono<String> addToCart(@PathVariable Long itemId, OrderRequest request, WebSession session) {
         return orderService
                 .updateOrder(itemId, request.getAction(), session.getId())
                 .thenReturn("redirect:/");
